requirements.txt
Pillow==9.4.0
black==22.12.0
flake8==5.0.4
isort[colors]==5.11.4
pytest==7.2.0
requests==2.28.1
urwid==2.1.2
